Unify ChunkSpec
